therapist
this prompt will make ChatGPT act as a good enough therapist for anyone in need. given, its not an amazing therapist but it'll get the job done for those who need it.
(i put it r ight next to the mental health advisor so its easier to find for anyone looking for it.